# Conflicts:
#	Next2/src/Next2/Services/Menu/MenuService.cs
#	Next2/src/Next2/ViewModels/Tablet/NewOrderViewModel.cs

+++ Next2/src/Next2/Models/CustomerModel.cs
@@ -43,7 +43,7 @@ namespace Next2.Models
 
         public int GiftCardCount { get; set; }
 
-        public float GiftCardTotal { get; set; }
+        public decimal GiftCardTotal { get; set; }
 
         public List<GiftCardModel> GiftCards { get; set; } = new();
